backend/requirements.txt
# Required Packages
flask
flask-cors
python-dotenv
readability-lxml
beautifulsoup4
openai
duckduckgo_search
requests